Individuals APIs Individual Details API         Individual Details Edit API         Individul Specific Contribution Add API         Individul Specific Contribution Edit API

+++ src/app/schema/forane_schema.py
@@ -1,4 +1,5 @@
 from dataclasses import Field
+from decimal import Decimal
 from common.schema import SchemaBase
 from pydantic import Field
 
@@ -12,7 +13,7 @@ class ForaneInfoSchemaBase(SchemaBase):
     forName: str | None = Field(default=None, description="Forane Church Name",alias="for_name")
     forLocation:str | None = Field(default=None,  description="Forane Location",alias="for_location")
     forVicarName:str | None  = Field(default=None, description="Forane Vicar Name",alias="for_vicar_name")
-    forTotalContribution:int | None  = Field(default=None, description="Forane Total Contribution",alias="for_total_contribution_amount")
+    forTotalContribution:Decimal | None  = Field(default=None, description="Forane Total Contribution",alias="for_total_contribution_amount")
     forContactNumber:str | None  = Field(default=None, description="Forane Contact Number",alias="for_contact_number")
 
     class Config:
